requirements.txt
# Core web framework
Flask==3.0.3

# Session management
Flask-Session==0.5.0

# Task queue
celery==5.4.0
redis==5.0.1

# Image processing + HEIF support
Pillow
pillow-heif

# Certs
cryptography

# System utilities
psutil==5.9.8

# Config loading (.env)
python-dotenv==1.0.1

# Flask deps (usually installed automatically, but safe to keep pinned)
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.3
